Small Testing Server
A small server to test out the book API.  all methods are in the
BooksAPI document.  Still a lot to implement once we decide how to store
sessions.  For now this can be used to test out html views and such.
Package.json contains some random crap that needs to be cleaned up for
the final version, leaving it there to test out stuff.